Move exceptions to exceptions file
